Add "Null" entry to `StructureTypes.xml`
A "Null" entry can be used to take up the slot at index `0`. That allows the `StructureType` index values to correspond with the `StrctureID` enum values. Potentially we can `static_cast` between `structureTypeIndex` values and `StructureID` enum values.

Additionally, having a null entry would allow us a place to list all available properties that can be set, and the default values of those properties. The defaults are currently hardcoded. Potentially the null entry could serve as a source for those defaults, or at least documentation as to what they are.

On a related note, we can potentially remove properties from `StructureType` entries when they are set to the default value. When the file was generated, all properties were always listed, even if they were default values for a field that `StructureType` didn't expect to use.