Set `Tube` and `AirShaft` as `IsSelfSustained`
The `Tube` and `AirShaft` won't trasmit a `CommandCenter` connection if they're not connected, though they also don't become disabled and highlighted in red when they are not connected. They are a rather passive structure, that are there when needed and connected.

The game hasn't been disabled unconnected tubes and highlighting them in red only because it doesn't currently do any processing for `Tube` structures. That `StructureClass` is never passed to `updateStructures` in `StructureManager`. This seems to be a bit of an oversight, which was perhaps not noticed because tubes don't really do anything, other than transmit a `CommandCenter` connection, and that is handled outside of regular `Structure` processing.