Reduce `BioWasteProcessingCapacity` to `120`
Early game needs `4` residences to support the initial population. With a processing capacity of `30` units per residence, and likely only a need for `4` residences, there's an effective limit of `120` units of waste processing early game. Maybe that should be the value used for balancing. By the time population grows to fill a fifth residence, maybe it's time to build a new recycling facility, or research some tech to improve the processing capacity.